# src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.0.0",
    "title": "Magalu Produtos API",
    "description": "Swagger da API Omni Channel"
  },
  "host": "localhost:3002",
  "basePath": "/",
  "tags": [
    {
      "name": "Produtos",
      "description": "Endpoints relacionados a produto;"
    },
    {
      "name": "Clientes",
      "description": "Endpoints relacionados a clientes;"
    },
    {
      "name": "Listas",
      "description": "Endpoints relacionados a listas;"
    },
    {
      "name": "Lojas",
      "description": "Endpoints relacionados a lojas;"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/produtos/": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "description": "Endpoint para obter uma lista de produtos.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Produto encontrado!",
            "schema": {
              "$ref": "#/definitions/Produtos"
            }
          },
          "400": {
            "description": "Desculpe, tivemos um problema com a requisição!"
          },
          "404": {
            "description": "Não existem produtos cadastrados!"
          }
        }
      }
    },
    "/lojas/": {
      "get": {
        "tags": [
          "Lojas"
        ],
        "description": "Endpoint para obter todas as lojas disponíveis.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lojas encontradas"
          },
          "400": {
            "description": "Desculpe, tivemos um problema com a requisição"
          },
          "404": {
            "description": "Lojas não encontradas"
          }
        }
      }
    },
    "/clientes/": {
      "post": {
        "tags": [
          "Clientes"
        ],
        "description": "Endpoint para criar um novo cliente.",
        "parameters": [
          {
            "name": "NovoCliente",
            "in": "body",
            "description": "Informações do cliente",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/NovoCliente"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Nome, CPF, RG, CEP e numero são campos obrigatórios!"
          }
        }
      }
    },
    "/listas/": {
      "post": {
        "tags": [
          "Listas"
        ],
        "description": "Endpoint para inserir produto em uma lista.",
        "parameters": [
          {
            "name": "AdicionarProduto",
            "in": "body",
            "description": "Os ids devem ser conhecidos!",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/AdicionarProduto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto inserido com sucesso!"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Todos os campos são obrigatórios!"
          }
        }
      },
      "delete": {
        "tags": [
          "Listas"
        ],
        "description": "Endpoint para inserir produto em uma lista.",
        "parameters": [
          {
            "name": "AdicionarProduto",
            "in": "body",
            "description": "Os ids devem ser conhecidos!",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/AdicionarProduto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto removido com sucesso!"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Todos os campos são obrigatórios!"
          }
        }
      }
    },
    "/listas/{clienteId}": {
      "get": {
        "tags": [
          "Listas"
        ],
        "description": "Endpoint para exibir todas as listas de um mesmo cliente.",
        "parameters": [
          {
            "name": "clienteId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Produto inserido com sucesso!",
            "schema": {
              "$ref": "#/definitions/ListarListas"
            }
          },
          "400": {
            "description": "Todos os campos são obrigatórios!"
          }
        }
      }
    },
    "/listas/{listaId}": {
      "put": {
        "tags": [
          "Listas"
        ],
        "description": "Endpoint para atualizar status de uma lista.",
        "parameters": [
          {
            "name": "listaId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "AtualizarStatus",
            "in": "body",
            "description": "Atualizar status de uma lista!",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/AtualizarStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Atualizado com sucesso!"
          },
          "400": {
            "description": "Desculpe, tivemos um problema com a requisição!"
          }
        }
      }
    }
  },
  "definitions": {
    "Produtos": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "produto": {
          "type": "string",
          "example": "Iphone SE"
        },
        "categoria": {
          "type": "string",
          "example": "celular"
        },
        "preco": {
          "type": "number",
          "example": 1000
        },
        "marca": {
          "type": "string",
          "example": "Apple"
        },
        "createdAt": {
          "type": "string",
          "example": "2021-09-12T18:34:44.057Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "2021-09-12T18:34:44.057Z"
        }
      }
    },
    "NovoCliente": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "example": "Cláudia"
        },
        "cpf": {
          "type": "string",
          "example": "381.493.542-04"
        },
        "rg": {
          "type": "string",
          "example": "000000"
        },
        "cep": {
          "type": "string",
          "example": "05036-001"
        },
        "cidade": {
          "type": "string",
          "example": "SP"
        },
        "estado": {
          "type": "string",
          "example": "SP"
        },
        "rua": {
          "type": "string",
          "example": "Navegantes"
        },
        "numero": {
          "type": "number",
          "example": 2016
        },
        "complemento": {
          "type": "string",
          "example": "Ap 101"
        },
        "createdAt": {
          "type": "string",
          "example": "2021-09-12T18:34:44.057Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "2021-09-12T18:34:44.057Z"
        }
      },
      "required": [
        "nome",
        "cpf",
        "rg",
        "cep",
        "numero"
      ]
    },
    "AdicionarProduto": {
      "type": "object",
      "properties": {
        "clienteId": {
          "type": "number",
          "example": 2
        },
        "lojaId": {
          "type": "number",
          "example": 2
        },
        "produtoId": {
          "type": "number",
          "example": 4
        }
      },
      "required": [
        "clienteId",
        "lojaId",
        "produtoId"
      ]
    },
    "ListarListas": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "status": {
          "type": "string",
          "example": "Em andamento"
        },
        "createdAt": {
          "type": "string",
          "example": "2021-09-12T18:36:09.553Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "2021-09-12T18:36:09.553Z"
        },
        "ClienteId": {
          "type": "number",
          "example": 1
        },
        "LojaId": {
          "type": "number",
          "example": 2
        }
      }
    },
    "AtualizarStatus": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "Em andamento"
        }
      }
    }
  }
}